requirements.txt
pyTelegramBotAPI==4.27.0
Flask==3.1.1
python-dotenv==1.0.1
google-generativeai==0.4.1
